`tests/test_reports.py`: Merge with `testing/test_reports.py`
